refactor: TIO-226: extract sound button into own widget
