+++ lib/pages/tuner/play_sound_page.dart
@@ -10,9 +10,9 @@ import 'package:tiomusic/pages/tuner/tuner_functions.dart';
 import 'package:tiomusic/src/rust/api/api.dart';
 import 'package:tiomusic/util/color_constants.dart';
 import 'package:tiomusic/util/constants.dart';
-import 'package:tiomusic/util/util_midi.dart';
 import 'package:tiomusic/widgets/dismiss_keyboard.dart';
 import 'package:tiomusic/widgets/input/number_input_and_slider_int.dart';
+import 'package:tiomusic/widgets/tuner/sound_button.dart';
 
 const double buttonWidth = 40;
 const double buttonPadding = 4;
@@ -169,76 +169,6 @@ class _PlaySoundPageState extends State<PlaySoundPage> {
           ],
         ),
       ),
-    );
-  }
-}
-
-class SoundButton extends StatefulWidget {
-  final int midiNumber;
-  final int idx;
-  final ActiveReferenceSoundButton buttonListener;
-
-  const SoundButton({super.key, required this.midiNumber, required this.idx, required this.buttonListener});
-
-  @override
-  State<SoundButton> createState() => _SoundButtonState();
-}
-
-class _SoundButtonState extends State<SoundButton> {
-  late double _concertPitch;
-
-  @override
-  void initState() {
-    super.initState();
-    _concertPitch = (Provider.of<ProjectBlock>(context, listen: false) as TunerBlock).chamberNoteHz;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return ListenableBuilder(
-      listenable: widget.buttonListener,
-      builder: (context, child) {
-        return Listener(
-          onPointerDown: (details) async {
-            setState(() {
-              if (widget.buttonListener.buttonOn) {
-                widget.buttonListener.turnOff();
-
-                if (widget.buttonListener.buttonIdx != widget.idx) {
-                  widget.buttonListener.turnOn(widget.idx, midiToFreq(widget.midiNumber, concertPitch: _concertPitch));
-                }
-              } else {
-                widget.buttonListener.turnOn(widget.idx, midiToFreq(widget.midiNumber, concertPitch: _concertPitch));
-              }
-            });
-          },
-          child: Padding(
-            padding: const EdgeInsets.all(buttonPadding),
-            child: Container(
-              width: buttonWidth,
-              height: 60,
-              decoration: BoxDecoration(
-                color:
-                    widget.buttonListener.buttonIdx == widget.idx && widget.buttonListener.buttonOn
-                        ? ColorTheme.primary
-                        : ColorTheme.surface,
-                borderRadius: BorderRadius.circular(20),
-              ),
-              child: Center(
-                child: Text(
-                  midiToNameOneChar(widget.midiNumber),
-                  style: TextStyle(
-                    color:
-                        widget.buttonListener.buttonIdx == widget.idx && widget.buttonListener.buttonOn
-                            ? ColorTheme.surface
-                            : ColorTheme.primary,
-                  ),
-                ),
-              ),
-            ),
-          ),
-        );
-      },
     );
   }
 }
